+++ src/main/kotlin/de/lancom/genesis/java/GenesisJavaPlugin.kt
@@ -4,9 +4,17 @@ import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.plugins.JavaPlugin
 
+private val javaPlugins = setOf(
+    "java-platform",
+    "java-library",
+    "java",
+)
+
 class GenesisJavaPlugin : Plugin<Project> {
     override fun apply(project: Project) {
-        project.pluginManager.apply(JavaPlugin::class.java)
+        if (!javaPlugins.any(project.pluginManager::hasPlugin)) {
+            project.pluginManager.apply(JavaPlugin::class.java)
+        }
         project.extensions.create("genesisJava", GenesisJavaExtension::class.java)
     }
 }