+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.1.3] - 2022-03-21
 ### Changed
 - fixed enableJava11
+- do not apply Java Plugin when other java plugins are already active
 
 ## [1.1.2] - 2022-03-18
 ### Changed